+++ tests/unit/visualize/tStem.m
@@ -1,0 +1,13 @@
+classdef tStem < MAGVisualizationTestCase & ColorSupportTestCase & MarkerSupportTestCase
+% TSTEM Unit tests for "mag.graphics.chart.Stem" class.
+
+    properties (Constant)
+        ClassName = "mag.graphics.chart.Stem"
+        GraphClassName = "matlab.graphics.chart.primitive.Stem"
+    end
+
+    properties (TestParameter)
+        Properties = {struct(Name = "LineStyle", Value = '-'), ...
+            struct(Name = "LineStyle", Value = '--')}
+    end
+end