+++ resources/ReleaseNotes.md
@@ -8,3 +8,4 @@
 - Rename `Result` to `PSD`
 - Remove support for `Filters` in charts
 - Fix issues with setting colors in charts
+- Add tests for area, scatter, stairs and stems plots